requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
scipy>=1.10.0
opencv-python>=4.8.0
scikit-image>=0.21.0
Pillow>=10.0.0
PyYAML>=6.0
tqdm>=4.66.0
matplotlib>=3.8.0
albumentations>=1.4.0
SimpleITK>=2.3.0
pandas>=2.0.0
